visualize() failed for batch_size more than 64
The code was always saving the images with in a 8x8 image frame. So it can hold maximum 64 images. That's why it was not failing if batch_size is less or equals to 64. If batch_size is more than 64, it was giving error.

As a solution we can increase the image frame dimensions as square. So that the batch size does not overflow in the image frame.