Fix: tensorboard scalar graph is overlapping with previous data when training is started from checkpoint
In case the training is loaded from checkpoint, the new training results will be overlapped by the old ones since the counter is not reinitalized by loading.

Solution:
load() is now returning with the counter when the checkpoint was saved. In case the load is successful the counter will be reinitialized